requirements.txt
ctransformers == 0.2.27
sentence-transformers == 2.2.2
pinecone-client
langchain == 0.2.1
flask
pypdf
python-dotenv
-e .